Refactor to 'package by annotation model'
