+++ src/main/java/com/brunodles/jsoupparser/exceptions/MissingSelectorException.java
@@ -1,7 +1,9 @@
 package com.brunodles.jsoupparser.exceptions;
 
+import com.brunodles.jsoupparser.bigannotation.CssSelector;
+
 /**
- * This exception is thrown when field is not annotated with {@link com.brunodles.jsoupparser.CssSelector}
+ * This exception is thrown when field is not annotated with {@link CssSelector}
  */
 public class MissingSelectorException extends RuntimeException {
     public MissingSelectorException(String methodName) {
