+++ src/main/java/com/brunodles/jsoupparser/smallanotation/annotations/Selector.java
@@ -1,9 +1,10 @@
-package com.brunodles.jsoupparser.annotations;
+package com.brunodles.jsoupparser.smallanotation.annotations;
 
 import java.lang.annotation.*;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-public @interface NestedCollector {
+public @interface Selector {
+    String value();
 }
