+++ src/main/java/com/brunodles/jsoupparser/smallanotation/annotations/TextCollector.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.annotations;
+package com.brunodles.jsoupparser.smallanotation.annotations;
 
 import java.lang.annotation.*;
