+++ src/test/java/com/brunodles/jsoupparser/bigannotation/NestedRootModel.java
@@ -1,8 +1,7 @@
 package com.brunodles.jsoupparser.bigannotation;
 
-import com.brunodles.jsoupparser.CssSelector;
-import com.brunodles.jsoupparser.collectors.NestedCollector;
-import com.brunodles.jsoupparser.collectors.TextElementCollector;
+import com.brunodles.jsoupparser.bigannotation.collectors.NestedCollector;
+import com.brunodles.jsoupparser.bigannotation.collectors.TextElementCollector;
 
 public interface NestedRootModel {
