+++ src/test/java/com/brunodles/jsoupparser/bigannotation/collectors/TextElementCollectorTest.java
@@ -1,4 +1,4 @@
-package com.brunodles.jsoupparser.collectors;
+package com.brunodles.jsoupparser.bigannotation.collectors;
 
 import com.brunodles.jsoupparser.JsoupParser;
 import com.brunodles.jsoupparser.doubles.UnknownException;
